Allow empty AWS_ENDPOINT env var (#33)
The AWS_ENDPOINT env var allows us to overwrite the default AWS base URL for localstack. But our default behaviour should be not to set an endpoint at all and use whatever the default in the AWS SDK is.

An empty/missing AWS_ENDPOINT will be handled as an empty string, and passing that to `Options.BaseEndpoint` is the behaviour that we want.

This will help the task start in real AWS environments.

For SSM-20 #patch