POC for accepting separate documents
## To allow documents to arrive separately

Remove the validation requiring instrument/application combinations in `set_validator.go`.

However, this should be replaced by validation to ensure that e.g. LP2 documents have no `CaseNo` because they create new cases, but LPA117 documents must have a `CaseNo`

## To support additional documents

For the extra documents we want to add quick support for, use a new `generic_parser` that doesn't need to have the XML structure mapped to Go. It is still validated against the XSD during parsing.

These documents are automatically sent on to Sirius, but we should replace this with an allowlist ([SSM-46](https://opgtransform.atlassian.net/browse/SSM-46)).