Validate embedded documents against their XSDs
Each embedded document is decoded and validated against its XSD.

In order to pass the individual errors onto the HTTP response, I've introduced a new Problem struct: this can be used to define an error message (and, optionally, additional validation errors) that can be exposed to the sender of the request.

This also enabled changing the response structure to match the legacy API so our suppliers can switch between the two more easily.

Fixes SSM-70 #minor