Add Sirius mock (#60)
Adds an imposter container that mocks Sirius's behaviour, which removes the requirement for an external network.

But integrating with Sirius is helpful, because it lets us do end-to-end tests and check mapping of the XML document, so I've added an additional `docker-compose.sirius.yml` file which uses (and requires) Sirius to be running locally.

Therefore:
- `make start` allows you to do a minimal local stand-up of the app
- `make start-sirius` requires Sirius to be running locally and lets you ingest documents fully

Fixes SSM-19 #minor