Add golangci-lint and gosec (#59)
* Add golangci-lint and gosec

* Fix linting issues

#patch

* Attach gosec to scanning network

So it can be run alone in CI

* Remove duplicate setup line

#patch

* Allow gosec to write sarif file

#patch

* Add ReadHeaderTimeout

Allow the server 5 seconds to read the request headers to mitigate against DDoS attacks that hang the connection open.

#patch

* Switch to local golangci-lint image

Allows us to generate CLI and SARIF output so we have reproducible results.

Requires a Dockerfile because libxml2 must be installed to install `lestrrat-go/libxml2` for linting.

#patch

* Halt pipeline if go-lint or gosec fails

Should not try to build and push image if linting issues were discovered.

#patch