Add integration test script
Starts up the application and sends example files to it to check the are correctly parsed.

For files which create cases, it also checks that a message has been added to the SQS queue.

I updated the Sirius mock to generate unique UIDs on each request, which is used to ensure the SQS message relates to the most recent file.

Fixed a few issues with our existing test files.

Removed LPA002 and LPA002R from `NewCaseDocuments`, as they should be supplied alongside LP2s which create the case.

Fixes SSM-72 #minor