Add Pact tests (#82)
* Add Pact tests

Rather than mocking the HTTP component, use a Pact mock server and generate a contract file.

Upload that file to the broker to ensure the contract can be met by Sirius (the provider).

Fixes SSM-89 #patch

* Rename `uid` -> `uId`

This is the actual field name in the API, it just worked before because `json.Unmarshal` falls back to working case-insensitively.

#patch