Explain error when case UID is invalid (#134)
If the case UID is malformed, Sirius will respond with a 400 error. This needs to be reported back to the client properly, rather than being treated as a generic error.

We use the same error status code (400) but a different error message, to provide a better hint at what needs to be fixed.

Fixes SSM-150 #minor