SSM-155 Return CaseNo when document has previously been processed (#153)
* Return CaseNo when document has previously been processed

* Fix formatting of UID in response

To match existing responses, 7000 UIDs should not have dashes in them

* Alter response of duplicate documents

Treat the request as a success and provide the UID in a separate data field.

* Allow retrying of failed files

If a document has been sent before but failed, try processing it again.

If a document is still processing, or completed successfully, the request will still be bounced back as a duplicate

---------

Co-authored-by: Greg Tyler <greg.tyler@digital.justice.gov.uk>